refactor: PhotoCard component
Transfer the methods referring to the PhotoCard component to the App
component
