+++ src/components/PageHeader.js
@@ -1,0 +1,5 @@
+import { ref, computed } from 'vue'
+
+export const data = ref("")
+
+export const textoAtualizado = computed(() => data.value)
